+++ src/nv_string.c
@@ -262,7 +262,7 @@ ptrdiff_t strViewNext(const StrView *sv, ptrdiff_t idx, UcdCP *outCP) {
     if (idx >= (ptrdiff_t)sv->len) {
         goto endReached;
     }
-    size_t runLen = ucdCh8RunLen(sv->buf[idx]);
+    uint8_t runLen = ucdCh8RunLen(sv->buf[idx]);
 
     if (runLen == 0 || idx + runLen > sv->len) {
         goto endReached;
@@ -293,7 +293,7 @@ ptrdiff_t strViewPrev(const StrView *sv, ptrdiff_t idx, UcdCP *outCP) {
     if (idx < 0) {
         goto endReached;
     }
-    size_t runLen = ucdCh8RunLen(sv->buf[idx]);
+    uint8_t runLen = ucdCh8RunLen(sv->buf[idx]);
     if (runLen == 0 || idx + runLen > sv->len) {
         goto endReached;
     }
